Organize Depends/Suggests/Imports
Putting pkgdown in depends loads it for all users when loading the package, but it's not needed by the users, so I put it in Suggests. I moved everything needed by code in the vignette to Imports. All will be installed anyways because of the `dependencies=TRUE` flag in Dockerfile.